Modifiqué Los métodos de Negocio para evitar el problema del cambio de tipo, cambiando de float a double y haciendo cambios de los métodos. También eliminé lo que había que hacía de test, y agregué 2 test nuevo que prueban algunos casos, tanto de precio de prendas como de ganancia total del negocio.

+++ tp1MacoWins/src/main/java/ar/edu/grupo5/jm/dss/macoWins/Procedencia.java
@@ -2,5 +2,5 @@ package ar.edu.grupo5.jm.dss.macoWins;
 
 public interface Procedencia {
 
-	public float factorMultiplicativo();	
+	public double factorMultiplicativo();	
 }
